Split some tcp code into separate files.
