+++ server/main.cpp
@@ -7,28 +7,10 @@
 #include <sys/socket.h>
 #include <netdb.h>
 #include <unistd.h> // for close
-
+#include "tcp.hpp"
 
 int main() {
-    struct protoent* protocol_entry = getprotobyname("tcp");
-    const int tcp_protocol_num = protocol_entry->p_proto;
-    
-    int server_socket = socket(AF_INET, SOCK_STREAM, tcp_protocol_num);
-    if (server_socket == -1) {
-        std::cerr << "Bad socket!\n";
-        exit(-1);
-    }
-
-    struct sockaddr_in addr;
-    addr.sin_family = AF_INET;
-    addr.sin_port = htons(7070);
-    addr.sin_addr.s_addr = INADDR_ANY;
-
-    int res = bind(server_socket, (struct sockaddr *)&addr, sizeof(addr));
-    if (res == -1) {
-        std::cerr << "Failed bind: " << strerror(errno) << "\n";
-        exit(-1);
-    }
+    int server_socket = create_server_socket(INADDR_ANY, 7070);
 
     listen(server_socket, 1);
