+++ server/tcp.hpp
@@ -1,0 +1,17 @@
+#ifndef TCP_H
+#define TCP_H
+
+#include <cstdint>
+#include <cstdio>
+#include <iostream>
+#include <iterator>
+#include <string.h>
+#include <netinet/in.h>
+#include <string>
+#include <sys/socket.h>
+#include <netdb.h>
+#include <unistd.h> // for close
+
+int create_server_socket(uint32_t addr, uint16_t port);
+
+#endif
